✨ Refactor client systems, engine and UI
