+++ client/src/systems/RenderSyncSystem.ts
@@ -108,7 +108,7 @@ export class RenderSyncSystem {
             const originalMaterial = (Array.isArray(baseMesh.material) ? baseMesh.material[0] : baseMesh.material) as THREE.MeshStandardMaterial;
             const material = originalMaterial.clone();
 
-            material.onBeforeCompile = (shader: THREE.Shader) => {
+            material.onBeforeCompile = (shader: { vertexShader: string; fragmentShader: string; uniforms: { [uniform: string]: THREE.IUniform } }) => {
                 shader.vertexShader = `
                     attribute float aPhaseShift;
                     varying float vPhase;
